+++ packages/author-profile/__tests__/mocks.js
@@ -1,31 +1,19 @@
-import { fixtureGenerator } from "@times-components/provider-test-tools";
-import longSummaryLength from "../author-profile-constants";
+jest.mock("@times-components/article-list", () =>
+  // eslint-disable-next-line global-require
+  require("./mock-article-list")
+);
+jest.mock("@times-components/gradient", () => "Gradient");
+jest.mock("@times-components/icons", () => ({
+  IconTwitter: "IconTwitter"
+}));
+jest.mock("@times-components/image", () => "Image");
+jest.mock("@times-components/link", () => ({
+  TextLink: "TextLink"
+}));
+jest.mock("@times-components/pagination", () => {
+  const id = x => x;
 
-export const pageSize = 3;
-export const mockArticles = fixtureGenerator.makeArticleMocks({
-  pageSize: 3,
-  withImages: true
+  return {
+    withPageState: id
+  };
 });
-export const mockArticlesWithoutImages = fixtureGenerator.makeArticleMocks({
-  longSummaryLength,
-  pageSize: 3,
-  withImages: false
-});
-export const mockAuthor = fixtureGenerator.makeAuthor({ withImages: true });
-export const mockAuthorWithoutImages = fixtureGenerator.makeAuthor({
-  withImages: false
-});
-
-export const props = {
-  adConfig: {},
-  analyticsStream: () => {},
-  author: mockAuthor,
-  fetchMore: () => {},
-  isLoading: false,
-  onArticlePress: () => {},
-  onTwitterLinkPress: () => {},
-  page: 1,
-  pageSize: 3,
-  refetch: () => {},
-  slug: "deborah-haynes"
-};