+++ packages/test-utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.6"></a>
+## [0.2.6](https://github.com/newsuk/times-components/compare/@times-components/test-utils@0.2.5...@times-components/test-utils@0.2.6) (2018-08-02)
+
+
+
+
+**Note:** Version bump only for package @times-components/test-utils
+
 <a name="0.2.5"></a>
 ## [0.2.5](https://github.com/newsuk/times-components/compare/@times-components/test-utils@0.2.4...@times-components/test-utils@0.2.5) (2018-07-31)