+++ packages/tracking/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.9"></a>
+## [2.0.9](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.0.8...@times-components/tracking@2.0.9) (2018-08-02)
+
+
+
+
+**Note:** Version bump only for package @times-components/tracking
+
 <a name="2.0.8"></a>
 ## [2.0.8](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.0.7...@times-components/tracking@2.0.8) (2018-07-31)